requirements.txt
fastapi
uvicorn
pymupdf
python-dotenv
python-multipart

sendgrid

crewai
langchain
langchain-openai

openai
tiktoken
